All these are part of our app.
My sister and I created an app for the MedHacks 2020 hackathon that helps the community fight addiction. When we exported our app, it created all these files, but if you want to access the code you look at the .pjx file from the Microsoft Visual FoxPro. Anothere method to view all the files is to unzip the app.zip file and view it in Microsoft Visual FoxPro.